chore: `CategoryList` 배열 -> `category.constant.ts` 파일로 이동
